fix: rename biaxial chart
