+++ apps/dapp/src/components/Charts/index.ts
@@ -1,7 +1,7 @@
 import LineChartComponent from './LineChart';
-import BiAxialLineChartComponent from './BiAxialLineChart';
+import BiAxialAreaChartComponent from './BiAxialAreaChart';
 import IntervalTogglerComponent from './IntervalToggler';
 
 export const LineChart = LineChartComponent;
-export const BiAxialLineChart = BiAxialLineChartComponent;
+export const BiAxialAreaChart = BiAxialAreaChartComponent;
 export const IntervalToggler = IntervalTogglerComponent;
